refactor: reorganize src/ into modules and add .gitignore
Restructure src/ directory into logical modules:
- src/core/      — logger, types, exit codes
- src/protocol/  — Thor protocol packet definitions
- src/usb/       — USB device communication layer
- src/firmware/   — TAR parsing, MD5 verification, LZ4 decompression

Update CMakeLists.txt source paths and include directories.
Update all #include directives to use module-relative paths.
Add .gitignore for build artifacts.
