+++ src/usb/usb_device.cpp
@@ -1,4 +1,4 @@
-#include "usb_device.h"
+#include "usb/usb_device.h"
 #include <iostream>
 #include <iomanip>
 #include <sstream>
